feat: replace in-memory db with postgres and typeorm in album module

+++ src/album/album.controller.ts
@@ -4,17 +4,18 @@ import {
   Delete,
   Get,
   HttpCode,
-  HttpStatus,
+  HttpStatus, NotFoundException,
   Param,
   ParseUUIDPipe,
   Post,
   Put,
   UsePipes,
-  ValidationPipe,
+  ValidationPipe
 } from '@nestjs/common';
 import { AlbumDto } from './dto/album.dto';
-import { Album } from './types/album';
 import { AlbumService } from './album.service';
+import Album from './entities/album.entity';
+import { ApiNotFoundResponse } from '@nestjs/swagger';
 
 @Controller('album')
 export class AlbumController {
@@ -27,14 +28,17 @@ export class AlbumController {
   }
 
   @Put(':id')
+  @ApiNotFoundResponse({ description: 'Album was not found.' })
   @UsePipes(new ValidationPipe({ skipNullProperties: true }))
   async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateAlbumDto: AlbumDto,
   ): Promise<Album> {
     const album = await this.albumService.findOne(id);
-    this.albumService.throwNotFoundException(album, id);
-    return this.albumService.update(id, album, updateAlbumDto);
+    if(!album) {
+      throw new NotFoundException('Album was not found.');
+    }
+    return this.albumService.update(id,updateAlbumDto);
   }
 
   @Get()
@@ -43,21 +47,27 @@ export class AlbumController {
   }
 
   @Get(':id')
+  @ApiNotFoundResponse({ description: 'Album was not found.' })
   async findOne(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
   ): Promise<Album> {
     const album = await this.albumService.findOne(id);
-    this.albumService.throwNotFoundException(album, id);
+    if(!album) {
+      throw new NotFoundException('Album was not found.');
+    }
     return album;
   }
 
   @Delete(':id')
+  @ApiNotFoundResponse({ description: 'Album was not found.' })
   @HttpCode(HttpStatus.NO_CONTENT)
   async remove(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
   ): Promise<void> {
     const album = await this.albumService.findOne(id);
-    this.albumService.throwNotFoundException(album, id);
+    if (!album) {
+      throw new NotFoundException('Album was not found.');
+    }
     await this.albumService.remove(album);
   }
 }
